Refresh cache by executing sources selected

Refresh cache by executing sources selected

Usage

sourcoise_refresh(
  what = NULL,
  force_exec = TRUE,
  unfreeze = TRUE,
  quiet = FALSE,
  init_fn = getOption("sourcoise.init_fn"),
  root = NULL,
  log = "INFO",
  .progress = TRUE
)

Arguments

what

(tibble) a tibble as generated by sourcoise_status(), possibly filtered, (defaut to source_status() )

force_exec

(boolean) if TRUE code is executed, no matter what is cached

unfreeze

(boolean) (default TRUE) when possible, unfreeze and uncache .qmd files in a quarto project when data used by those .qmd has been refreshed

quiet

no message if TRUE

init_fn

(function) (default NULL) execute a function

root

(default NULL) force root to be set, instead of letting the function finding the root, for advanced uses

log

(string) (default "INFO") log levels as in logger::log_threshold() (c("OFF", "INFO", ...)), comes with a performance cost

.progress

(boolean) (default TRUE) displays a progression bar based on previous execution timings

Value

a list of r scripts (characters) executed, with timing and success

See also